fix: move dali preprocessor to custom handler
